Fix bug in firmware info retrieving
The m_firmware_info variable was uninitialized, which could lead to not
actually requesting the firmware info from the hardware (leading to a
random version number). By explicitly setting the version numbers to zero,
this request is forced. This also ensures that the correct configuration
is loaded, as the version number is used for deciding which
configuration to use.