Reorganize package structure
This package will grow with the inclusion and additions of the icl_comm
and icl_core libraries and possible replacements. A new package
structure makes sense here.
We also properly separate header files and implementation files.

We now have
- `control` for the control and core driver related things
- `serial` for the serial interface and related functionality
- Other files stay top-level for now

We only move the files without modifying them to keep the Git blame
functionality.