Switch to modern CMake
This is now a modern CMake package.
We use CMake 3.10, which is Bionic's (Ubuntu 18.04) shipping version.

The package's primary target is `schunk_svh` (`libschunk_svh.so`) that
will be the core library for ROS1 and ROS2 drivers.
Other libraries are meant to stay internal.

There are still two system dependencies:
- `icl_comm` (for `icl_comm_websocket`)
- and on `icl_core`.
Both will be included or replaced in later commits.