Include `icl_comm_websocket` into this package
The functionalities of `icl_comm_websocket` were specifically developed
for the Schunk SVH driver. We now get them back into this repository and
drop the dependency of `icl_comm`.

The files' content is not changed except for adjustments to the
includes, full name qualifications of some types, and the
websocket-specific preprocessor directives.

Jsoncpp is currently included in cmake with `PckConfig` and zmq seems to
ship with latest Ubuntu versions. See
[here](https://packages.ubuntu.com/search?keywords=libzmq)

Let's see if this robustly works on *Bionic* and *Focal* once we have
the CI pipeline running.