Remodel the `SVHReceiveThread` with std::thread
The `SVHReceiveThread` is now a *normal* class, which is instantiated as a
member in client code. Users then call its `run()` method in a separate
thread and its `stop()` method before joining with the main thread.
This is cleaner and easier to understand than the previous approach of
inheriting from a thread class.

We also replace time measurement with `std::chrono` features.